+++ projects/_01_ingest/cqc_api/jobs/delta_clean_cqc_location_data.py
@@ -1,6 +1,5 @@
 import os
 import sys
-import warnings
 
 os.environ["SPARK_VERSION"] = "3.5"
 
@@ -35,7 +34,6 @@ from utils.column_names.raw_data_files.cqc_location_api_columns import (
 )
 from utils.column_values.categorical_column_values import (
     CareHome,
-    LocationType,
     PrimaryServiceType,
     RegistrationStatus,
     RelatedLocation,
@@ -116,8 +114,8 @@ def main(
         cqc_location_df, known_la_providerids
     )
 
-    cqc_location_df = remove_non_social_care_locations(cqc_location_df)
     cqc_location_df = remove_records_from_locations_data(cqc_location_df)
+
     cqc_location_df = utils.format_date_fields(
         cqc_location_df,
         date_column_identifier=CQCLClean.registration_date,  # This will format both registration date and deregistration date
@@ -128,24 +126,21 @@ def main(
     )
 
     cqc_location_df = impute_historic_relationships(cqc_location_df)
-    registered_locations_df = select_registered_locations_only(cqc_location_df)
 
-    dimension_update_date = registered_locations_df.agg(
-        F.max(Keys.import_date)
-    ).collect()[0][0]
+    dimension_update_date = cqc_location_df.agg(F.max(Keys.import_date)).collect()[0][0]
 
     # Create regulated activity dimension
     regulated_activity_delta = create_dimension_from_missing_struct_column(
-        registered_locations_df,
+        cqc_location_df,
         CQCL.regulated_activities,
         regulated_activities_destination,
         dimension_update_date,
     )
-    registered_locations_df = registered_locations_df.drop(CQCL.regulated_activities)
+    cqc_location_df = cqc_location_df.drop(CQCL.regulated_activities)
 
-    registered_locations_df, regulated_activity_delta = (
+    cqc_location_df, regulated_activity_delta = (
         remove_locations_that_never_had_regulated_activities(
-            registered_locations_df, regulated_activity_delta
+            cqc_location_df, regulated_activity_delta
         )
     )
     regulated_activity_delta = extract_registered_manager_names(
@@ -161,12 +156,12 @@ def main(
 
     # Create specialisms dimension
     specialisms_delta = create_dimension_from_missing_struct_column(
-        registered_locations_df,
+        cqc_location_df,
         CQCL.specialisms,
         specialisms_destination,
         dimension_update_date,
     )
-    registered_locations_df = registered_locations_df.drop(CQCL.specialisms)
+    cqc_location_df = cqc_location_df.drop(CQCL.specialisms)
 
     specialisms_delta = extract_from_struct(
         specialisms_delta,
@@ -195,15 +190,13 @@ def main(
 
     # Create GAC service dimension
     gac_service_delta = create_dimension_from_missing_struct_column(
-        registered_locations_df,
+        cqc_location_df,
         CQCL.gac_service_types,
         gac_service_destination,
         dimension_update_date,
     )
     # Drop care home from registered location df - stored in gac service dimension
-    registered_locations_df = registered_locations_df.drop(
-        CQCL.gac_service_types, CQCL.care_home
-    )
+    cqc_location_df = cqc_location_df.drop(CQCL.gac_service_types, CQCL.care_home)
 
     gac_service_delta = extract_from_struct(
         gac_service_delta,
@@ -211,9 +204,6 @@ def main(
         CQCLClean.services_offered,
     )
 
-    registered_locations_df, gac_service_delta = remove_specialist_colleges(
-        registered_locations_df, gac_service_delta
-    )
     gac_service_delta = allocate_primary_service_type(gac_service_delta)
 
     gac_service_delta = realign_carehome_column_with_primary_service(
@@ -228,19 +218,17 @@ def main(
     )
 
     # Final cleaning on fact table
-    registered_locations_df = add_related_location_column(registered_locations_df)
+    cqc_location_df = add_related_location_column(cqc_location_df)
 
     # Create postcode matching dimension
     postcode_matching_delta = create_postcode_matching_dimension(
-        registered_locations_df,
+        cqc_location_df,
         ons_postcode_directory_df,
         postcode_matching_destination,
         dimension_update_date,
     )
 
-    registered_locations_df = registered_locations_df.drop(
-        CQCL.postal_code, CQCL.postal_address_line1
-    )
+    cqc_location_df = cqc_location_df.drop(CQCL.postal_code, CQCL.postal_address_line1)
 
     utils.write_to_parquet(
         postcode_matching_delta,
@@ -250,7 +238,7 @@ def main(
     )
 
     utils.write_to_parquet(
-        registered_locations_df,
+        cqc_location_df,
         cleaned_cqc_location_destination,
         mode="overwrite",
         partitionKeys=cqcPartitionKeys,
@@ -504,10 +492,6 @@ def impute_missing_registration_dates(df: DataFrame) -> DataFrame:
         ).otherwise(df[CQCLClean.imputed_registration_date]),
     )
     return df
-
-
-def remove_non_social_care_locations(df: DataFrame) -> DataFrame:
-    return df.where(df[CQCL.type] == LocationType.social_care_identifier)
 
 
 def impute_historic_relationships(df: DataFrame) -> DataFrame:
@@ -796,66 +780,6 @@ def add_related_location_column(df: DataFrame) -> DataFrame:
         ),
     )
     return df
-
-
-def remove_specialist_colleges(
-    cqc_df: DataFrame, gac_services_dimension: DataFrame
-) -> tuple[DataFrame, DataFrame]:
-    """
-    Removes rows where 'Specialist college service' is the only service listed in 'services_offered'.
-
-    We do not include locations which are only specialist colleges in our
-    estimates. This function identifies and removes the ones listed in the locations dataset.
-
-    Args:
-        cqc_df (DataFrame): A dataframe without services_offered, but where the location_ids need to be aligned
-        gac_services_dimension (DataFrame): A cleaned locations dataframe with the services_offered column already created.
-
-    Returns:
-        tuple[DataFrame, DataFrame]: cqq_df, gac_services_dimension with locations which are only specialist colleges removed.
-    """
-    # The below just prevents IDEs from warning you that the "Column object is not callable"
-    # - this is a false warning and does not cause an error
-    # noinspection PyCallingNonCallable
-    to_remove = gac_services_dimension.where(
-        (
-            gac_services_dimension[CQCLClean.services_offered][0]
-            == Services.specialist_college_service
-        )
-        & (F.size(gac_services_dimension[CQCLClean.services_offered]) == 1)
-        & (gac_services_dimension[CQCLClean.services_offered].isNotNull())
-    ).select(CQCLClean.location_id, Keys.import_date)
-
-    cqc_df = cqc_df.join(
-        to_remove,
-        on=[CQCLClean.location_id, Keys.import_date],
-        how="left_anti",
-    )
-
-    gac_services_dimension = gac_services_dimension.join(
-        to_remove,
-        on=[CQCLClean.location_id, Keys.import_date],
-        how="left_anti",
-    )
-
-    return cqc_df, gac_services_dimension
-
-
-def select_registered_locations_only(locations_df: DataFrame) -> DataFrame:
-    invalid_rows = locations_df.where(
-        (locations_df[CQCL.registration_status] != RegistrationStatus.registered)
-        & (locations_df[CQCL.registration_status] != RegistrationStatus.deregistered)
-    ).count()
-
-    if invalid_rows != 0:
-        warnings.warn(
-            f"{invalid_rows} row(s) has/have an invalid registration status and have been dropped."
-        )
-
-    locations_df = locations_df.where(
-        locations_df[CQCL.registration_status] == RegistrationStatus.registered
-    )
-    return locations_df
 
 
 def add_cqc_sector_column_to_cqc_locations_dataframe(